+++ src/index.ts
@@ -4,6 +4,5 @@ export * from './tx-builder/types';
 export * from './tx-builder/types/GovernanceMethodTypes';
 export * from './tx-builder/types/GovernanceReturnTypes';
 export * from './tx-builder/types/GovDelegationMethodTypes';
-export { DydxStateChecker } from './tx-builder/services/DydxStateChecker';
 
 export * from './tx-builder/config';
